Apply the improvements given by Tamer.
In this commit, improvements were made to the system, such as
considering the naming of private fields and methods. Additionally,
the code was cleaned up significantly by removing unused code, such
as the 'IncreaseQuantity' and 'DecreaseQuantity' methods from the
'Product' class. Another enhancement was focusing on code reusability.
For instance, a generic method named 'IsValid' was created in the
'Product' class to handle the values of '_quantity' and '_price'.
Furthermore, a generic method called 'CheckIfNumeric' was introduced
in the 'Menu' class to replace both 'CheckIfDouble' and 'CheckIfInt'
methods.